Refactor getJson() and getPage to post() and get()
