+++ src/fi/bitrite/android/ws/api/RestClient.java
@@ -6,6 +6,7 @@ import android.widget.Toast;
 import fi.bitrite.android.ws.R;
 import fi.bitrite.android.ws.auth.NoAccountException;
 import fi.bitrite.android.ws.auth.http.HttpAuthenticationFailedException;
+import fi.bitrite.android.ws.auth.http.HttpAuthenticator;
 import fi.bitrite.android.ws.auth.http.HttpSessionContainer;
 import fi.bitrite.android.ws.util.Tools;
 import fi.bitrite.android.ws.util.http.HttpException;
@@ -16,22 +17,69 @@ import org.apache.http.HttpStatus;
 import org.apache.http.NameValuePair;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.entity.UrlEncodedFormEntity;
+import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.protocol.HttpContext;
 import org.apache.http.util.EntityUtils;
-
 import java.io.IOException;
 import java.util.List;
 
 /**
- * Base class for classes that interface with the REST API via POST.
- * see https://github.com/rfay/Warmshowers.org/wiki/Warmshowers-RESTful-Services-for-Mobile-Apps
+ * Base class for classes that use GET to either scrape the WS website for information
+ * or interface with the REST API.
  */
-public class RestClient extends HttpReader {
+public class RestClient {
 
-    private static final String TAG = "RestClient";
+    boolean authenticationPerformed;
 
-    protected String getJson(String url, List<NameValuePair> params) throws HttpException, IOException {
+    public RestClient() {
+        setAuthenticationPerformed(false);
+    }
+
+    public boolean isAuthenticationPerformed() {
+        return authenticationPerformed;
+    }
+
+    private void setAuthenticationPerformed(boolean authenticationPerformed) {
+        this.authenticationPerformed = authenticationPerformed;
+    }
+
+    protected String get(String simpleUrl) throws HttpException {
+        HttpClient client = HttpUtils.getDefaultClient();
+        String json;
+        int responseCode;
+
+        try {
+            String url = HttpUtils.encodeUrl(simpleUrl);
+            HttpGet get = new HttpGet(url);
+            HttpContext context = HttpSessionContainer.INSTANCE.getSessionContext();
+
+            HttpResponse response = client.execute(get, context);
+            HttpEntity entity = response.getEntity();
+            responseCode = response.getStatusLine().getStatusCode();
+
+            json = EntityUtils.toString(entity, "UTF-8");
+
+            if (responseCode == HttpStatus.SC_FORBIDDEN ||
+                    responseCode == HttpStatus.SC_UNAUTHORIZED) {
+                if (!isAuthenticationPerformed()) {
+                    authenticate();
+                    json = get(simpleUrl);  // TODO: Remove ugly and unnecessary recursion
+                } else {
+                    throw new HttpException("Couldn't authenticate user");
+                }
+            }
+
+        } catch (Exception e) {
+            throw new HttpException(e);
+        } finally {
+            client.getConnectionManager().shutdown();
+        }
+
+        return json;
+    }
+
+    protected String post(String url, List<NameValuePair> params) throws HttpException, IOException {
         HttpClient client = HttpUtils.getDefaultClient();
         String json = "";
 
@@ -47,7 +95,7 @@ public class RestClient extends HttpReader {
                     responseCode == HttpStatus.SC_UNAUTHORIZED) {
                 if (!isAuthenticationPerformed()) {
                     authenticate();
-                    return getJson(url, params);
+                    return post(url, params);
                 } else {
                     throw new HttpException("Couldn't authenticate user");
                 }
@@ -87,4 +135,11 @@ public class RestClient extends HttpReader {
         }
         return;
     }
+
+    protected void authenticate() throws NoAccountException, IOException {
+        HttpAuthenticator authenticator = new HttpAuthenticator();
+        authenticator.authenticate();
+        setAuthenticationPerformed(true);
+    }
+
 }
