+++ src/fi/bitrite/android/ws/host/impl/HttpHostFeedback.java
@@ -1,6 +1,6 @@
 package fi.bitrite.android.ws.host.impl;
 
-import fi.bitrite.android.ws.api.HttpReader;
+import fi.bitrite.android.ws.api.RestClient;
 import fi.bitrite.android.ws.model.Feedback;
 import fi.bitrite.android.ws.util.GlobalInfo;
 
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 /**
  * Retrieves feedback for a given host.
  */
-public class HttpHostFeedback extends HttpReader {
+public class HttpHostFeedback extends RestClient {
 
     /**
      * Given the ID of a WarmShowers user, retrieve feedback about him.
@@ -20,7 +20,7 @@ public class HttpHostFeedback extends HttpReader {
     public ArrayList<Feedback> getFeedback(int id) {
         String simpleUrl = new StringBuilder().append(GlobalInfo.warmshowersBaseUrl).append("/user/")
                 .append(id).append("/json_recommendations").toString();
-        String json = getPage(simpleUrl);
+        String json = get(simpleUrl);
 
         FeedbackJsonParser parser = new FeedbackJsonParser(json);
         return parser.getFeedback();
