+++ src/fi/bitrite/android/ws/host/impl/RestTextSearch.java
@@ -1,13 +1,10 @@
 package fi.bitrite.android.ws.host.impl;
 
-import android.util.Log;
-
 import org.apache.http.NameValuePair;
 import org.apache.http.message.BasicNameValuePair;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import fi.bitrite.android.ws.WSAndroidApplication;
 import fi.bitrite.android.ws.api.RestClient;
 import fi.bitrite.android.ws.host.Search;
 import fi.bitrite.android.ws.model.HostBriefInfo;
@@ -15,7 +12,6 @@ import fi.bitrite.android.ws.util.GlobalInfo;
 import fi.bitrite.android.ws.util.http.HttpException;
 
 import java.io.IOException;
-import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -43,7 +39,7 @@ public class RestTextSearch extends RestClient implements Search {
         List<NameValuePair> args = new ArrayList<NameValuePair>();
         args.add(new BasicNameValuePair("keyword", this.keyword));
 
-        String json = getJson(WARMSHOWERS_HOST_BY_KEYWORD_URL, args);
+        String json = post(WARMSHOWERS_HOST_BY_KEYWORD_URL, args);
         JSONObject allJson = new JSONObject(json);
 
         List<HostBriefInfo> list = new ArrayList<HostBriefInfo>();
